% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_sub.R
\name{add_sub}
\alias{add_sub}
\title{Adds annotation underneath a plot, similar to the `sub` parameter in base R.}
\usage{
add_sub(plot, label, x = 0.5, y = 0.5, hjust = 0.5, vjust = 0.5,
  vpadding = grid::unit(1, "lines"), font_family = "",
  font_face = "plain", colour = "black", size = 14, angle = 0,
  lineheight = 0.9)
}
\arguments{
\item{plot}{A ggplot object or gtable object derived from a ggplot object.}

\item{label}{The label with which the plot should be annotated. Can be a plotmath expression.}

\item{x}{The x position of the label}

\item{y}{The y position of the label}

\item{hjust}{Horizontal justification}

\item{vjust}{Vertical justification}

\item{vpadding}{Vertical padding. The total vertical space added to the label, given in grid
units. By default, this is added equally above and below the label. However, by changing the
y and vjust parameters, this can be changed.}

\item{font_family}{The font family}

\item{font_face}{The font face ("plain", "bold", etc.)}

\item{colour}{Text color}

\item{size}{Point size of text}

\item{angle}{Angle at which text is drawn}

\item{lineheight}{Line height of text}
}
\value{
A gtable object holding the modified plot.
}
\description{
Adds annotation underneath a plot, similar to the `sub` parameter in base R.
}
\examples{
p1 <- ggplot(mtcars, aes(mpg, disp)) + geom_line(colour = "blue") + background_grid(minor='none')
ggdraw(add_sub(p1, "This is an annotation.\\nAnnotations can span multiple lines."))

# You can also do this repeatedly. Just be aware that new annotations are added above previous ones:
p2 <- add_sub(p1, expression(paste(a^2+b^2, " = ", c^2)))
p3 <- add_sub(p2, "This formula has no relevance here:", y  = 0, vjust = 0)
ggdraw(p3)

#This code also works with faceted plots:
plot.iris <- ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point() + facet_grid(. ~ Species) + stat_smooth(method = "lm") +
  background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  panel_border() # and a border around each panel
p2 <- add_sub(plot.iris, "Annotation underneath a faceted plot, left justified.", x = 0, hjust = 0)
ggdraw(p2)

# Finally, it is possible to move the annotation inside of the plot if desired.
ggdraw(add_sub(p1, "Annotation inside plot", vpadding=grid::unit(0, "lines"), y = 6, x = 0.02, hjust = 0))
}

