% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_canvas.R
\name{axis_canvas}
\alias{axis_canvas}
\title{Generates a canvas onto which one can draw axis-like objects.}
\usage{
axis_canvas(plot, axis = "y", xlim = NULL, ylim = NULL)
}
\description{
Generates a canvas onto which one can draw axis-like objects.
}
\examples{
# continuous case
x <- seq(0, 10, .1)
d <- data.frame(x,
                linear = x,
                squared = x*x/5,
                cubed = x*x*x/25) \%>\%
  gather(fun, y, -x)

pmain <- ggplot(d, aes(x, y, group = fun)) + geom_line()  +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(limits = c(0, 60))

paxis <- axis_canvas(pmain, axis = "y") +
  geom_text(data = filter(d, x==max(x)), aes(y = y, label=paste0(" ", fun)),
            x = 0, hjust = 0, vjust = 0.5)
ggdraw(insert_yaxis_grob(pmain, get_panel(paxis), grid::unit(.25, "null")))
}
