% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowplot.R
\docType{package}
\name{cowplot-package}
\alias{cowplot}
\alias{cowplot-package}
\title{Streamlined Plot Theme and Plot Annotations for 'ggplot2'}
\description{
Provides various features that help with creating publication-quality
figures with 'ggplot2', such as a set of themes, functions to align
plots and arrange them into complex compound figures, and functions
that make it easy to annotate plots and or mix plots with images. The
package was originally written for internal use in the Wilke lab,
hence the name (Claus O. Wilke's plot package). It has also been used
extensively in the book Fundamentals of Data Visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wilkelab.org/cowplot/}
  \item Report bugs at \url{https://github.com/wilkelab/cowplot/issues}
}

}
\author{
\strong{Maintainer}: Claus O. Wilke \email{wilke@austin.utexas.edu} (\href{https://orcid.org/0000-0002-7470-9261}{ORCID})

}
\keyword{internal}
