% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_grob}
\alias{draw_grob}
\title{Draw a grob.}
\usage{
draw_grob(grob, x = 0, y = 0, width = 1, height = 1)
}
\arguments{
\item{grob}{The grob to place.}

\item{x}{The x location of the lower left corner of the grob.}

\item{y}{The y location of the lower left corner of the grob.}

\item{width}{Width of the grob.}

\item{height}{Height of the grob.}
}
\description{
Places an arbitrary grob somewhere onto the drawing canvas. By default, coordinates run from
0 to 1, and the point (0, 0) is in the lower left corner of the canvas.
}

