% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw_plot}
\alias{draw_plot}
\title{Draw a (sub)plot.}
\usage{
draw_plot(plot, x = 0, y = 0, width = 1, height = 1)
}
\arguments{
\item{plot}{The plot to place}

\item{x}{x location of the lower left corner of the plot}

\item{y}{y location of the lower left corner of the plot}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\description{
Places a plot somewhere onto the drawing canvas. By default, coordinates run from
0 to 1, and the point (0, 0) is in the lower left corner of the canvas.
}

