% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_vector_image}
\alias{draw_vector_image}
\title{Draw a vector image}
\usage{
draw_vector_image(image, x = 0, y = 0, width = 1, height = 1,
  scale = 1, clip = "inherit", .gridSVG = FALSE)
}
\arguments{
\item{image}{The image to place. Can be a file path, or an image previously created by \code{grImport2::readPicture()}.}

\item{x}{The x location of the lower left corner of the image.}

\item{y}{The y location of the lower left corner of the image.}

\item{width}{Width of the image.}

\item{height}{Height of the image.}

\item{scale}{Scales the image relative to the rectangle defined by \code{x}, \code{y}, \code{width}, \code{height}. A setting
of \code{scale = 1} indicates no scaling.}

\item{clip}{Set to "on" to clip the image relative to the box into which it is draw (useful for \code{scale > 1}).
Note that clipping doesn't always work as expected, due to limitations of the grid graphics system.}
}
\description{
Places a vector image somewhere onto the drawing canvas. By default, coordinates run from
0 to 1, and the point (0, 0) is in the lower left corner of the canvas. Requires the \code{grImport2} and \code{grConvert}
package to work, and fails gracefully if that package is not installed.
}
\examples{
library(ggplot2)
library(cowplot)
cairoFile <- system.file("SVG", "lwd-rsvg.svg", package="grImport2")
otherFile <- system.file("SVG", "lwd.svg", package="grImport2")

ggdraw() + draw_vector_image(grImport2::readPicture(cairoFile))
dev.off()
ggdraw() + draw_vector_image(cairoFile)
dev.off()
ggdraw() + draw_vector_image(otherFile)
dev.off()

plot_grid(NULL,
          ggdraw() + draw_vector_image(otherFile),
          NULL,
          NULL,
          ncol=2, nrow=2, labels="AUTO")
}
