% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_component.R
\name{get_plot_component}
\alias{get_plot_component}
\alias{plot_component_names}
\alias{plot_components}
\title{Get plot components}
\usage{
get_plot_component(plot, pattern)

plot_component_names(plot)

plot_components(plot)
}
\arguments{
\item{plot}{a ggplot or gtable}

\item{pattern}{the name of the component}
}
\value{
a grob
}
\description{
Get plot components
}
