% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axes.R
\name{get_y_axis}
\alias{get_y_axis}
\alias{get_x_axis}
\title{Get plot axes}
\usage{
get_y_axis(plot, side = c("l", "r"))

get_x_axis(plot, side = c("b", "t"))
}
\arguments{
\item{plot}{a ggplot or gtable}

\item{side}{which side of the plot?}
}
\description{
Get plot axes
}
