% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggsave}
\alias{ggsave}
\title{Cowplot reimplementation of ggsave.}
\usage{
ggsave(filename = NULL, plot = ggplot2::last_plot(),
  device = default_device(filename), path = NULL, scale = 1,
  width = par("din")[1], height = par("din")[2], units = c("in", "cm",
  "mm"), dpi = 300, limitsize = TRUE, ...)
}
\description{
This function should behave just like \code{ggsave()} from ggplot2,
with the main difference being that by default it doesn't use the Dingbats
font for pdf output. If you ever have trouble with this function, you can
try using \code{ggplot2::ggsave()} instead.
}

