% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.R
\name{gtable_remove_grobs}
\alias{gtable_remove_grobs}
\title{Remove named elements from gtable}
\usage{
gtable_remove_grobs(table, names, ...)
}
\arguments{
\item{table}{The table from which grobs should be removed}

\item{names}{A character vector of the grob names (as listed in \code{table$layout})
that should be removed}

\item{...}{Other parameters passed through to \code{gtable_filter}.}
}
\description{
Remove named elements from gtable
}
