% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.R
\name{gtable_squash_cols}
\alias{gtable_squash_cols}
\title{Set the width of given colums to 0.}
\usage{
gtable_squash_cols(table, cols)
}
\arguments{
\item{table}{The gtable on which to operate}

\item{cols}{Numerical vector indicating the columns whose width should be set to zero.}
}
\description{
Set the width of given colums to 0.
}
