% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.R
\name{gtable_squash_rows}
\alias{gtable_squash_rows}
\title{Set the height of given rows to 0.}
\usage{
gtable_squash_rows(table, rows)
}
\arguments{
\item{table}{The gtable on which to operate}

\item{rows}{Numerical vector indicating the rows whose heights should be set to zero.}
}
\description{
Set the height of given rows to 0.
}
