% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_component.R
\name{is_zeroGrob}
\alias{is_zeroGrob}
\title{is zero grob}
\usage{
is_zeroGrob(x)
}
\arguments{
\item{x}{A grob}
}
\value{
A logical value
}
\description{
Function to check if a grob is a zero grob
}
