% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_gtable.R
\name{plot_to_gtable}
\alias{plot_to_gtable}
\title{Convert plot or other graphics object into a gtable}
\usage{
plot_to_gtable(plot)
}
\arguments{
\item{plot}{The plot or other graphics object to convert into a gtable. Here, \code{plot} can be an
object of the following classes: \code{\link[ggplot2]{ggplot}}, \code{\link[grDevices:recordPlot]{recordedplot}},
\code{\link[grid]{grob}}, or \code{\link[gtable]{gtable}}. Alternatively, \code{plot} can be
a function creating a plot when called (see examples for \code{\link[=plot_grid]{plot_grid()}}).}
}
\description{
This function does it's best attempt to take whatever you provide it and turn it into a gtable.
It is primarily meant to convert ggplot plots into gtables, but it will also take any grid
object (grob), a recorded R base plot, or a function that generates an R base plot.
}
