% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_null_device.R
\name{png_null_device}
\alias{png_null_device}
\alias{pdf_null_device}
\alias{cairo_null_device}
\alias{agg_null_device}
\title{Null devices}
\usage{
png_null_device(width, height)

pdf_null_device(width, height)

cairo_null_device(width, height)

agg_null_device(width, height)
}
\arguments{
\item{width}{Device width in inch}

\item{height}{Device height in inch}
}
\description{
Null devices to be used when rendering graphics in the background. See
\code{\link[=set_null_device]{set_null_device()}} for details.
}
