% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save_plot}
\alias{save_plot}
\title{Alternative to ggsave.}
\usage{
save_plot(filename, plot, cols = 1, rows = 1, base_width = 5,
  base_height = 3.5, dpi = 300, units = "in", useDingbats = FALSE)
}
\description{
This function is just a thin wrapper around ggsave.
}

