% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_cow}
\alias{theme_cow}
\title{Default cowplot theme}
\usage{
theme_cow(base_size = 14)
}
\arguments{
\item{base_size}{The overall font size. Default is 14.}
}
\value{
The theme.
}
\description{
After loading the cowplot package, this theme will be the default
for all graphs made with ggplot2.
}
\examples{
qplot(1:10, 1:10^2) + theme_cow(base_size=16)
}

