% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_nothing}
\alias{theme_nothing}
\title{Create a completely empty theme}
\usage{
theme_nothing(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{The overall font size. Default is 14.}
}
\value{
The theme.
}
\description{
The theme created by this function shows nothing but the plot panel. Unfortunately,
due to ggplot2 peculiarities, a little bit of padding remains to the left and bottom of
the panel. This can be removed by adding labs(x = NULL, y = NULL) to the plot, see examples.
}
\examples{
qplot(1:10, 1:10^2) + theme_cow(base_size = 16)
}

