% Generated by roxygen2 (4.0.2): do not edit by hand
\name{background_grid}
\alias{background_grid}
\title{Add/modify/remove the background grid in a ggplot2 plot}
\usage{
background_grid(major = c("xy", "x", "y", "only_minor", "none"),
  minor = c("xy", "x", "y", "none"), size.major = 0.2, size.minor = 0.5,
  colour.major = "grey90", colour.minor = "grey98")
}
\description{
This function provides a simple way to modify the background grid in ggplot2. It
doesn't do anything that can't be done just the same with theme(). However, it simplifies
creation of the most commonly needed variations.
}

