% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw_line}
\alias{draw_line}
\title{Draw a line.}
\usage{
draw_line(x, y, ...)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}

\item{...}{style parameters, such as \code{colour}, \code{alpha}, \code{size}, etc.}
}
\description{
This is a convenience function. It's just a thin wrapper around \code{geom_line()}.
}

