% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_plot}
\alias{draw_plot}
\title{Draw a (sub)plot.}
\usage{
draw_plot(plot, x = 0, y = 0, width = 1, height = 1)
}
\arguments{
\item{plot}{The plot to place. Can be a ggplot2 plot, an arbitrary gtable,
or a recorded base-R plot, as in \code{\link[=plot_grid]{plot_grid()}}.}

\item{x}{The x location of the lower left corner of the plot.}

\item{y}{The y location of the lower left corner of the plot.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\description{
Places a plot somewhere onto the drawing canvas. By default, coordinates run from
0 to 1, and the point (0, 0) is in the lower left corner of the canvas.
}
