% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_plot}
\alias{draw_plot}
\title{Draw a (sub)plot.}
\usage{
draw_plot(plot, x = 0, y = 0, width = 1, height = 1, scale = 1)
}
\arguments{
\item{plot}{The plot to place. Can be a ggplot2 plot, an arbitrary grob or gtable,
or a recorded base-R plot, as in \code{\link[=plot_to_gtable]{plot_to_gtable()}}.}

\item{x}{The x location of the lower left corner of the plot.}

\item{y}{The y location of the lower left corner of the plot.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}

\item{scale}{Scales the grob relative to the rectangle defined by \code{x}, \code{y}, \code{width}, \code{height}. A setting
of \code{scale = 1} indicates no scaling.}
}
\description{
Places a plot somewhere onto the drawing canvas. By default, coordinates run from
0 to 1, and the point (0, 0) is in the lower left corner of the canvas.
}
\examples{
# make a plot
p <- qplot(1:10, 1:10)
# draw into the top-right corner of a larger plot area
ggdraw() + draw_plot(p, .6, .6, .4, .4)
}
