% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw_plot_label}
\alias{draw_plot_label}
\title{Draw plot label}
\usage{
draw_plot_label(label, x = 0, y = 1, hjust = -0.3, vjust = 1.3,
  size = 14, fontface = "bold", ...)
}
\description{
This function adds a plot label to the upper left corner of a graph. It takes all the same parameters
as draw_text(), but has defaults that make it convenient to label graphs.
}

