% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_text}
\alias{draw_text}
\title{Draw text.}
\usage{
draw_text(text, x = 0.5, y = 0.5, size = 14, ...)
}
\arguments{
\item{text}{Character or expression vector specifying the text to be written.}

\item{x}{Vector of x coordinates.}

\item{y}{Vector of y coordinates.}

\item{size}{Font size of the text to be drawn.}

\item{...}{Style parameters, such as \code{colour}, \code{alpha}, \code{angle}, \code{size}, etc.}
}
\description{
This is a convenience function to plot multiple pieces of text at the same time. It cannot
handle mathematical expressions, though. For those, use \code{draw_label}.
}
\details{
Note that font sizes get scaled by a factor of 2.85, so sizes given here agree with font sizes used in
the theme. This is different from \code{geom_text} in ggplot2.

By default, the x and y coordinates specify the center of the text box. Set \code{hjust = 0, vjust = 0} to specify
the lower left corner, and other values of \code{hjust} and \code{vjust} for any other relative location you want to
specify.
}
\examples{
ggdraw() + draw_text("Hello World!")
}
