% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panel.R
\name{get_panel}
\alias{get_panel}
\title{Retrieve the panel of a plot}
\usage{
get_panel(plot)
}
\arguments{
\item{plot}{A ggplot or gtable from which to retrieve the panel}
}
\value{
A gtable object holding just the panel
}
\description{
This function extracts just the main panel from a ggplot. It only works for plots with exactly
one panel (i.e., plots that are not faceted).
}
