% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axes.R
\name{get_y_axis}
\alias{get_y_axis}
\alias{get_x_axis}
\title{Get plot Axes}
\usage{
get_y_axis(plot, side = c("l", "r"))

get_x_axis(plot, side = c("b", "t"))
}
\arguments{
\item{plot}{A ggplot or gtable.}

\item{side}{Which side of the plot is the axis on?}
}
\description{
These functions extract just the axes from a ggplot. \code{get_y_axis()} pulls
the y-axis, while \code{get_x_axis()} pulls the x-axis.
}
\examples{
library(ggplot2)

p <- ggplot(mpg, aes(displ, cty)) +
  geom_point()

ggdraw(get_y_axis(p))
ggdraw(get_x_axis(p + scale_x_continuous(position = "top"), side = "t"))

}
