% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw.R
\name{ggdraw}
\alias{ggdraw}
\title{Set up a drawing layer on top of a ggplot}
\usage{
ggdraw(plot = NULL, xlim = c(0, 1), ylim = c(0, 1))
}
\arguments{
\item{plot}{The plot to use as a starting point. Can be either a ggplot2 plot or an arbitrary gtable.}

\item{xlim}{The x-axis limits for the drawing layer (default is [0, 1]).}

\item{ylim}{The y-axis limits for the drawing layer (default is [0, 1]).}
}
\description{
Set up a drawing layer on top of a ggplot
}

