% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw.R
\name{ggdraw}
\alias{ggdraw}
\title{Set up a drawing layer on top of a ggplot}
\usage{
ggdraw(plot = NULL)
}
\arguments{
\item{plot}{The ggplot2 plot object to use as a starting point.}
}
\description{
Set up a drawing layer on top of a ggplot
}

