% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel_border}
\alias{panel_border}
\title{Add/remove the panel border in a ggplot2 plot}
\usage{
panel_border(colour = "gray80", size = 0.5, linetype = 1,
  remove = FALSE)
}
\description{
This function provides a simple way to modify the panel border in ggplot2. It
doesn't do anything that can't be done just the same with theme(). However, it
saves some typing.
}

