% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{panel_border}
\alias{panel_border}
\title{Add/remove the panel border in a ggplot2 plot}
\usage{
panel_border(colour = "gray80", size = 0.5, linetype = 1,
  remove = FALSE)
}
\arguments{
\item{colour}{The color of the border.}

\item{size}{Size.}

\item{linetype}{Line type.}

\item{remove}{If \code{TRUE}, removes the current panel border.}
}
\description{
This function provides a simple way to modify the panel border in ggplot2. It
doesn't do anything that can't be done just the same with \code{theme()}. However, it
saves some typing.
}
