% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_cowplot}
\alias{theme_cowplot}
\title{Create the default cowplot theme}
\usage{
theme_cowplot(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{Overall font size. Default is 14.}

\item{font_family}{Default font family.}

\item{line_size}{Default line size.}
}
\value{
The theme.
}
\description{
After loading the cowplot package, this theme will be the default
for all graphs made with ggplot2.
}
\examples{
qplot(1:10, (1:10)^2) + theme_cowplot(font_size = 15)
}
