% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_nothing}
\alias{theme_nothing}
\title{Create a completely empty theme}
\usage{
theme_nothing(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Overall font size. Default is 14.}

\item{base_family}{Base font family.}
}
\value{
The theme.
}
\description{
The theme created by this function shows nothing but the plot panel. Unfortunately,
due to ggplot2 peculiarities, a little bit of padding remains to the left and bottom of
the panel. This can be removed by adding \code{labs(x = NULL, y = NULL)} to the plot, see examples.
}
\examples{
# plot with small amount of remaining padding
qplot(1:10, (1:10)^2) + theme_nothing()
# remaining padding removed
qplot(1:10, (1:10)^2) + theme_nothing() + labs(x = NULL, y = NULL)
}

