% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_nothing}
\alias{theme_nothing}
\title{Create a completely empty theme}
\usage{
theme_nothing(font_size = 14, font_family = "", rel_small = 12/14)
}
\arguments{
\item{font_size}{Overall font size. Default is 14.}

\item{font_family}{Base font family.}

\item{rel_small}{Relative size of small text}
}
\value{
The theme.
}
\description{
The theme created by this function shows nothing but the plot panel.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(disp, mpg, color = cyl)) +
  geom_point() +
  theme_nothing()
}
