% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_nothing}
\alias{theme_nothing}
\title{Create a completely empty theme}
\usage{
theme_nothing(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{Overall font size. Default is 14.}

\item{base_family}{Base font family.}
}
\value{
The theme.
}
\description{
The theme created by this function shows nothing but the plot panel.
}
\examples{
qplot(1:10, (1:10)^2) + theme_nothing()
}
